net/usb: rndis_host:  Add Reset Resume
Add proper Reset Resume and Manager Power so that the
Host can take an RNDIS connection into the USB U3 State.

The Reset Resume is needed in case of certain failures.

QCOM will be patching at some point but for now this is
an early version of the patch.

Change-Id: I94b390d3ffb67a82efdb6255df9006a8bd4167ce
Signed-off-by: Kenny Kessler <kenny.kessler@motorola.com>
Reviewed-on: https://gerrit.mot.com/1307021
SLTApproved: Slta Waiver
SME-Granted: SME Approvals Granted
Tested-by: Jira Key
Reviewed-by: Kenneth Kessler <kennykessler@motorola.com>
Submit-Approved: Jira Key